requirements.txt
numpy>=1.23
pyyaml>=6.0
mpi4py>=3.1; platform_system != "Windows"
